pyproject.toml
[tool.poetry]
name = "stlib-restapi"
version = "0.1.0"
description = ""
authors = ["2018008068 <aza1200@naver.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "stlib_restapi"}]

[tool.poetry.dependencies]
python = "^3.9"
black = "^22.10.0"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"
pyjwt = "^2.6.0"
django-environ = "^0.9.0"
dj-database-url = "^1.0.0"
psycopg2-binary = "^2.9.5"
whitenoise = {extras = ["brotli"], version = "^6.2.0"}
gunicorn = "^20.1.0"
setuptools = "^65.6.3"
django = "4.0"


[build-system]
requires = ["poetry-core","setuptools"]
build-backend = "poetry.core.masonry.api"
